requirements.txt
pandas
openpyxl
psycopg2-binary
sqlalchemy
jupyter
nbformat
matplotlib
seaborn